+++ link_preview/CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 0.0.2
+
+* Added [CacheExtension] to provide convenient access to a [CacheManager] instance.
+
+## 0.0.3
+
+- Added [Cache] class to provide a global singleton instance of [CacheManager].
+- Added [CacheExtension.cacheManager] method to provide convenient access to the global singleton [Cache].